+++ changelog.md
@@ -1,6 +1,6 @@
 # grinder.dart changes
 
-## UNRELEASED
+## 0.7.1 (2015/5/19)
 - `Dart.run` now takes an optional `vmArgs`, a list of arguments passed to the Dart VM.
 - Added `downgrade` method do `Pub`.
 - Introduce `RunOptions` to support all `Process.run`/`Process.start`